+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-# :arrows_clockwise: Fortinet Fortigate As Built Report Changelog
+# :arrows_clockwise: Fortinet FortiGate As Built Report Changelog
 
 ## [0.1.0] - 2022-07-31
 
-- Initial Release with support of Forticare, System, Route and Firewall
+- Initial Release with support of FortiCare, System, Route and Firewall